refactor(editor): move batch executor into package
